refactor banuser
