+++ src/main/java/it/blackhat/symposium/managers/AdminManager.java
@@ -22,14 +22,4 @@ public interface AdminManager {
     Optional<Admin> findAdmin(String username, String password) throws SQLException;
 
 
-    /**
-     * Ban an user from the platform
-     *
-     * @param time  the duration of the ban
-     * @param email the e-mail of the user
-     * @return true if the operation go well, false otherwise
-     * @throws SQLException if db error occured
-     */
-    int banUser(Date time, String email) throws SQLException;
-
 }
