+++ src/main/java/it/blackhat/symposium/managers/AdminModelManager.java
@@ -25,11 +25,4 @@ public class AdminModelManager extends ConnectionManager implements AdminManager
                 new BeanHandler<>(AdminModel.class), username, password);
         return Optional.ofNullable(foundAdmin);
     }
-
-    @Override
-    public int banUser(Date time, String email) throws SQLException {
-        QueryRunner run = new QueryRunner(this.dataSource);
-        int rowMod = run.update(BAN, time, email);
-        return rowMod;
-    }
 }
