+++ src/main/java/it/blackhat/symposium/queries/AdminQuery.java
@@ -13,10 +13,7 @@ public final class AdminQuery {
     private AdminQuery() {
         super();
     }
-
     
     public static final String SIGN_IN = "SELECT * FROM admin "
             + "WHERE email=? AND password=?";
-    
-    public static final String BAN = "UPDATE user SET banLastDate=?";
 }
